// src/api/generated/model/pagedListMgmtDistributionSet.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * hawkBit REST APIs
 * Eclipse hawkBit™ is a domain-independent back-end framework for rolling out software updates to constrained edge devices as well as more powerful controllers and gateways connected to IP based networking infrastructure.

 * OpenAPI spec version: v1
 */
import type { MgmtDistributionSet } from './mgmtDistributionSet';
import type { Links } from './links';

export interface PagedListMgmtDistributionSet {
  content: MgmtDistributionSet[];
  total?: number;
  size?: number;
  _links?: Links;
}