// src/api/generated/model/pagedListMgmtDistributionSetType.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * hawkBit REST APIs
 * Eclipse hawkBit™ is a domain-independent back-end framework for rolling out software updates to constrained edge devices as well as more powerful controllers and gateways connected to IP based networking infrastructure.

 * OpenAPI spec version: v1
 */
import type { MgmtDistributionSetType } from './mgmtDistributionSetType';
import type { Links } from './links';

export interface PagedListMgmtDistributionSetType {
  content: MgmtDistributionSetType[];
  total?: number;
  size?: number;
  _links?: Links;
}